chore[react-devtools]: extract some utils into separate modules to unify implementations
